[250819] Day10 - 과제제출
