+++ diary/src/types/index.d.ts
@@ -1,0 +1,15 @@
+export type TComment = {
+  id: number;
+  text: string;
+  date: string;
+};
+
+export type TCard = {
+  id: number;
+  imgUrl: string;
+  state: string;
+  date: string;
+  title: string;
+  content: string;
+  comments: Comment[];
+};
