+++ diary/src/utils/utils.ts
@@ -1,0 +1,5 @@
+export const setTime = new Promise((resolve) => {
+  setTimeout(() => {
+    resolve('Time is up!');
+  }, 3000);
+});
